Refactor: move ArticleList, Article component to sectionTabs folder
