--- src/components/ArticleList/index.js
@@ -1,4 +1,0 @@
-import Article from './Article';
-import ArticleList from './ArticleList';
-
-export { Article, ArticleList };
