+++ src/components/SectionTabs/index.js
@@ -1,3 +1,5 @@
 import SectionTabList from './SectionTabList';
+import ArticleList from './ArticleList';
+import Article from './Article';
 
-export { SectionTabList };
+export { SectionTabList, ArticleList, Article };
